requirements.txt
dicttoxml==1.7.4

Flask==1.1.1
Flask-Admin
Flask-Bcrypt
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-MySQLdb==0.2.0
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==2.4.3
Flask-Uploads==0.2.1
Flask-WTF==0.14.3

psycopg2==2.8.6
psycopg2-binary==2.8.6

python-dotenv

requests==2.24.0

xmltodict==0.12.0
